% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_distr.R
\name{draw_zoabeta}
\alias{draw_zoabeta}
\title{Draw a zero-one-augmented beta distribution based on parameters or data}
\usage{
draw_zoabeta(x, draw_hist = FALSE)
}
\arguments{
\item{x}{Either named list or numeric vector of distribution parameters, or
a numeric vector of values in the interval \code{[0,1]} to which a
distribution can be fitted.}

\item{draw_hist}{(logical, default \code{FALSE}) In the case of \code{x}
being a numeric vector of data values drawn from a ZOA-beta distribution,
setting \code{hist=TRUE} will display a histogram for the values in
\code{(0,1)} rather than a fitted density curve. Ignored if \code{x} is a
named list or numeric vector of parameters.}
}
\value{
A ggplot object.
}
\description{
This function displays a ZOA-beta distribution as a ggplot object. It accepts
either a named vector or list of distribution parameters, or a vector of
numeric values in the interval [0,1]. If \code{x} is named vector or list of
parameters must include elements 'shape1' and 'shape2', while elements
'pzero' and 'pone' are optional and, if missing, will be set to zero.  If
\code{x} is a numeric vector of data values in the interval \code{[0,1]}, the
function \code{\link{fit_zoabeta}} will be called to estimate the
distribution parameters. The distribution will be displayed as a combination
of vertical lines indicating the probability of values of 0 and 1 (if
present), together with a density curve for values strictly between 0 and 1.
When \code{x} is a data vector, there is also the option of displaying the
values strictly between 0 and 1 as a histrogram rather than a density curve.
}
