% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoabeta_functions.R
\name{dzoabeta}
\alias{dzoabeta}
\alias{pzoabeta}
\alias{qzoabeta}
\alias{rzoabeta}
\title{Zero-one-augmented beta Distribution}
\usage{
dzoabeta(x, shape1, shape2, pzero = 0, pone = 0, log = FALSE)

pzoabeta(
  x,
  shape1,
  shape2,
  pzero = 0,
  pone = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

qzoabeta(
  p,
  shape1,
  shape2,
  pzero = 0,
  pone = 0,
  lower.tail = TRUE,
  log.p = FALSE
)

rzoabeta(n, shape1, shape2, pzero = 0, pone = 0)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{shape1, shape2}{Non-negative parameters of the beta distribution for
values greater than zero and less than one.}

\item{pzero}{Probability of a value of zero (default is 0).}

\item{pone}{Probability of a value of one (default is 0).}

\item{log, log.p}{If \code{TRUE}, log density / probability values are returned.}

\item{lower.tail}{If \code{TRUE} (default), probabilities are
\eqn{P[X \le x]}, otherwise \eqn{P[X \gt x]}.}

\item{p}{Vector of probabilities.}

\item{n}{Number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}
}
\value{
\code{dzoabeta} returns the density, \code{pzoabeta} returns the
  probability, \code{qzoabeta} returns the quantile, and \code{rzoabeta}
  generates random values.
}
\description{
Density, distribution function, quantile function and random generation for
the Zero-one-augmented beta distribution with parameters
\code{shape1}, \code{shape2}, \code{pzero} and \code{pone}.
}
