% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_distr.R
\name{fit_zoabeta}
\alias{fit_zoabeta}
\title{Maximum likelihood fitting of the zero-inflated beta distribution}
\usage{
fit_zoabeta(x, edge_threshold = 0.99)
}
\arguments{
\item{x}{Numeric vector of values between 0 and 1}

\item{edge_threshold}{(single numeric value; default 0.99) The maximum
proportion of zero and/or one values in \code{x} above which fitting
parameters for the beta component will be skipped.}
}
\value{
Named numeric vector with elements:
  \item{pzero}{Proportion of zero values}
  \item{pone}{Proportion of one values}
  \item{shape1}{Shape1 parameter of beta distribution (NA if not fitted)}
  \item{shape2}{Shape2 parameter of beta distribution (NA if not fitted)}
}
\description{
Fits a zero-inflated beta distribution to a vector of data values that lie in
the interval \code{[0,1]}.
}
